MAINT: refactor openstack_identity
Remove class - use functions instead. Now openstack_project

we now take `conn` - openstack connection object for each function

remove unnecessary functions
